request

+++ src/main/java/com/tma/tt/api/jpa/AreaJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.Area;
 
-public interface AreaJpaRepository extends CommonJpaRepository<Area, Integer> {
+public interface AreaJpaRepository extends CommonJpaRepository<Area, Long> {
 }
