+++ src/main/java/com/tma/tt/api/jpa/CsvCandidateJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.CsvCandidate;
 
-public interface CsvCandidateJpaRepository extends CommonJpaRepository<CsvCandidate, Integer> {
+public interface CsvCandidateJpaRepository extends CommonJpaRepository<CsvCandidate, Long> {
 }
