+++ src/main/java/com/tma/tt/api/jpa/DeliveryCenterJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.DeliveryCenter;
 
-public interface DeliveryCenterJpaRepository extends CommonJpaRepository<DeliveryCenter, Integer> {
+public interface DeliveryCenterJpaRepository extends CommonJpaRepository<DeliveryCenter, Long> {
 }
