+++ src/main/java/com/tma/tt/api/jpa/DeliveryGroupJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.DeliveryGroup;
 
-public interface DeliveryGroupJpaRepository extends CommonJpaRepository<DeliveryGroup, Integer> {
+public interface DeliveryGroupJpaRepository extends CommonJpaRepository<DeliveryGroup, Long> {
 }
