+++ src/main/java/com/tma/tt/api/jpa/QuestionChoiceJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.QuestionChoice;
 
-public interface QuestionChoiceJpaRepository extends CommonJpaRepository<QuestionChoice, Integer> {
+public interface QuestionChoiceJpaRepository extends CommonJpaRepository<QuestionChoice, Long> {
 }
