+++ src/main/java/com/tma/tt/api/jpa/ScheduleJpaRepository.java
@@ -4,5 +4,5 @@ import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.Area;
 import com.tma.tt.api.model.Schedule;
 
-public interface ScheduleJpaRepository extends CommonJpaRepository<Schedule, Integer> {
+public interface ScheduleJpaRepository extends CommonJpaRepository<Schedule, Long> {
 }
