+++ src/main/java/com/tma/tt/api/jpa/SubjectJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.Subject;
 
-public interface SubjectJpaRepository extends CommonJpaRepository<Subject, Integer> {
+public interface SubjectJpaRepository extends CommonJpaRepository<Subject, Long> {
 }
