+++ src/main/java/com/tma/tt/api/jpa/TaskAssignmentJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.TaskAssignment;
 
-public interface TaskAssignmentJpaRepository extends CommonJpaRepository<TaskAssignment, Integer> {
+public interface TaskAssignmentJpaRepository extends CommonJpaRepository<TaskAssignment, Long> {
 }
