+++ src/main/java/com/tma/tt/api/jpa/TaskCategoryJpaRepository.java
@@ -3,5 +3,5 @@ package com.tma.tt.api.jpa;
 import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.TaskCategory;
 
-public interface TaskCategoryJpaRepository extends CommonJpaRepository<TaskCategory, Integer> {
+public interface TaskCategoryJpaRepository extends CommonJpaRepository<TaskCategory, Long> {
 }
