+++ src/main/java/com/tma/tt/api/jpa/UserMentorJpaRepository.java
@@ -4,6 +4,6 @@ import com.tma.tt.api.common.CommonJpaRepository;
 import com.tma.tt.api.model.UserMentor;
 import com.tma.tt.api.model.UserTest;
 
-public interface UserMentorJpaRepository extends CommonJpaRepository<UserMentor, Integer> {
+public interface UserMentorJpaRepository extends CommonJpaRepository<UserMentor, Long> {
 }
