+++ src/main/java/com/tma/tt/api/repository/AreaRepository.java
@@ -4,6 +4,6 @@ import com.tma.tt.api.model.Area;
 
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface AreaRepository extends ResourceRepositoryV2<Area, Integer> {
+public interface AreaRepository extends ResourceRepositoryV2<Area, Long> {
 
 }
