+++ src/main/java/com/tma/tt/api/repository/DeliveryCenterRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.DeliveryCenter;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface DeliveryCenterRepository extends ResourceRepositoryV2<DeliveryCenter, Integer> {
+public interface DeliveryCenterRepository extends ResourceRepositoryV2<DeliveryCenter, Long> {
 
 }
