+++ src/main/java/com/tma/tt/api/repository/QuestionChoiceRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.QuestionChoice;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface QuestionChoiceRepository extends ResourceRepositoryV2<QuestionChoice, Integer> {
+public interface QuestionChoiceRepository extends ResourceRepositoryV2<QuestionChoice, Long> {
 
 }
