+++ src/main/java/com/tma/tt/api/repository/SkillRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.Skill;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface SkillRepository extends ResourceRepositoryV2<Skill, Integer> {
+public interface SkillRepository extends ResourceRepositoryV2<Skill, Long> {
 
 }
