+++ src/main/java/com/tma/tt/api/repository/SubjectRepository.java
@@ -4,6 +4,6 @@ import com.tma.tt.api.model.Subject;
 
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface SubjectRepository extends ResourceRepositoryV2<Subject, Integer> {
+public interface SubjectRepository extends ResourceRepositoryV2<Subject, Long> {
 
 }
