+++ src/main/java/com/tma/tt/api/repository/TaskCategoryRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.TaskCategory;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface TaskCategoryRepository extends ResourceRepositoryV2<TaskCategory, Integer> {
+public interface TaskCategoryRepository extends ResourceRepositoryV2<TaskCategory, Long> {
 
 }
