+++ src/main/java/com/tma/tt/api/repository/TaskTemplateRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.TaskTemplate;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface TaskTemplateRepository extends ResourceRepositoryV2<TaskTemplate, Integer> {
+public interface TaskTemplateRepository extends ResourceRepositoryV2<TaskTemplate, Long> {
 
 }
