+++ src/main/java/com/tma/tt/api/repository/TestRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.Test;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface TestRepository extends ResourceRepositoryV2<Test, Integer> {
+public interface TestRepository extends ResourceRepositoryV2<Test, Long> {
 
 }
