+++ src/main/java/com/tma/tt/api/repository/UserFeedbackRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.UserFeedback;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface UserFeedbackRepository extends ResourceRepositoryV2<UserFeedback, Integer> {
+public interface UserFeedbackRepository extends ResourceRepositoryV2<UserFeedback, Long> {
 
 }
