+++ src/main/java/com/tma/tt/api/repository/UserMentorRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.UserMentor;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface UserMentorRepository extends ResourceRepositoryV2<UserMentor, Integer> {
+public interface UserMentorRepository extends ResourceRepositoryV2<UserMentor, Long> {
 
 }
