+++ src/main/java/com/tma/tt/api/repository/UserTestResultRepository.java
@@ -3,6 +3,6 @@ package com.tma.tt.api.repository;
 import com.tma.tt.api.model.UserTestResult;
 import io.katharsis.repository.ResourceRepositoryV2;
 
-public interface UserTestResultRepository extends ResourceRepositoryV2<UserTestResult, Integer> {
+public interface UserTestResultRepository extends ResourceRepositoryV2<UserTestResult, Long> {
 
 }
